Add CT and fix two bugs at once
The first one: the third nested child in 'test.json' was being
considered 'grandchild', that is to say, stack pop was not
correctly managed. The problem arised when poping after a
nested previous array. Fixed with new control booleans.

Also, a problem to finish a parent node xml tag arised when
we had two consecutive nested levels. This is ensured and
tested with new 'test2.json'.